Make the token helper public
